requirements.txt
opencv-python-headless
matplotlib
numpy
pandas
Pillow
torch
torchvision
streamlit
streamlit_folium
openpyxl
plotly
gdown
